break(utils): move `HEX, toHEX, viaHEX` to `buffer` module
