+++ src/crypto.ts
@@ -1,4 +1,5 @@
-import { encode, toHEX } from 'worktop/utils';
+import { encode } from 'worktop/utils';
+import { toHEX } from 'worktop/buffer';
 import type { Algorithms, TypedArray } from 'worktop/crypto';
 
 export function digest(algo: Algorithms.Digest, message: string): Promise<string> {
