+++ src/utils.d.ts
@@ -13,28 +13,6 @@ declare global {
 
 
 /**
- * All 256 hexadecimal pairs
- * @NOTE Maximum index is `255`
- */
-export const HEX: readonly string[];
-
-
-/**
- * Convert an `ArrayBuffer` to a hexadecimal string.
- * @param {ArrayBuffer} input
- */
-export function toHEX(input: ArrayBuffer): string;
-
-
-/**
- * Decode a hexadecimal string into an `Uint8Array` instance.
- * @NOTE Pass output through `decode()` for string conversion.
- * @param {string} input
- */
-export function viaHEX(input: string): Uint8Array;
-
-
-/**
  * Generate a unique string of `len` length.
  * @NOTE Relies on `crypto` to produce cryptographically secure (CSPRNG) values.
  * @param {number} [len] The desired length (defaults to `11`)
