+++ src/utils.test.ts
@@ -7,88 +7,6 @@ import isUUID from 'is-uuid';
 
 // ---
 
-const HEX = suite('HEX');
-
-HEX('should be an Array', () => {
-	assert.instance(utils.HEX, Array);
-});
-
-HEX('should have 256 values', () => {
-	assert.is(utils.HEX.length, 256);
-});
-
-HEX('should have correct values', () => {
-	assert.is(utils.HEX[0], '00');
-	assert.is(utils.HEX[10], '0a');
-	assert.is(utils.HEX[255], 'ff');
-	assert.is(utils.HEX[256], undefined);
-
-	for (let i = 0; i < 256; i++) {
-		assert.is(utils.HEX[i], i.toString(16).padStart(2, '0'));
-	}
-});
-
-HEX.run();
-
-// ---
-
-const toHEX = suite('toHEX');
-
-toHEX('should be a function', () => {
-	assert.type(utils.toHEX, 'function');
-});
-
-toHEX('should return a string', () => {
-	const input = new Uint8Array;
-	assert.type(utils.toHEX(input), 'string');
-	assert.type(utils.toHEX(input.buffer), 'string');
-});
-
-toHEX('should convert input HEX string', () => {
-	const input = utils.encode('hello world');
-	assert.is(utils.toHEX(input), '68656c6c6f20776f726c64');
-	assert.is(utils.toHEX(input.buffer), '68656c6c6f20776f726c64');
-});
-
-toHEX.run();
-
-// ---
-
-const viaHEX = suite('viaHEX');
-
-viaHEX('should be a function', () => {
-	assert.type(utils.viaHEX, 'function');
-});
-
-viaHEX('should return a `Uint8Array` instance', () => {
-	assert.instance(utils.viaHEX(''), Uint8Array);
-	assert.instance(utils.viaHEX('input'), Uint8Array);
-});
-
-viaHEX('should decode HEX string', () => {
-	const expects = utils.encode('hello world');
-	const output = utils.viaHEX('68656c6c6f20776f726c64');
-	assert.equal(output, expects);
-
-	assert.is(
-		utils.decode(output),
-		'hello world'
-	);
-});
-
-viaHEX('should handle emoji content', () => {
-	const output = utils.viaHEX('54686520717569636b2062726f776e20f09fa68a206a756d7073206f766572203133206c617a7920f09f90b62e');
-
-	assert.is(
-		utils.decode(output),
-		'The quick brown 🦊 jumps over 13 lazy 🐶.'
-	);
-});
-
-viaHEX.run();
-
-// ---
-
 // https://github.com/lukeed/uid/blob/master/test/secure.js
 const uid = suite('uid');
